Chirag/requirements.txt
flask==2.2.3
requests==2.28.2
python-dotenv==1.0.0
pandas==2.0.1
matplotlib==3.7.1
pytest==7.3.1
